requirements-lock.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
defusedxml==0.7.1
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
markupsafe==2.1.5
more-itertools==10.3.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.38.0
openapi-core==0.19.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.1
parse==1.20.2
pathable==0.4.3
prance==23.6.21.0
pybars4==0.9.13
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
pymeta3==0.5.1
python-dotenv==1.0.1
pyyaml==6.0.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.19.1
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
ruff==0.5.2
semantic-kernel==1.3.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
yarl==1.9.4
